# Conflicts:
#	blackboard2/target/classes/com/example/demo/ServiceImp/AttendanceServiceImp.class

+++ blackboard2/src/main/java/com/example/demo/Dao/AnnouncementDao.java
@@ -5,6 +5,6 @@ import com.example.demo.Domain.Announcement;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface AnnouncementDao extends JpaRepository<Announcement, Integer> {
+public interface AnnouncementDao extends JpaRepository<Announcement, Long> {
 
 }
